Sets ruby version to 3.0
Ruby version 3.1 causes failures on Windows. This appears to be caused by the `childprocess` gem, and the way it launches executables. The error shows up when the Cucumber/Aruba library using `childprocess` to launch a program. `childprocess` is throwing an exception when it attempts to get a OS file handle for the file that output should be redirected to. The function call is returning -1, but the Windows "last error" isn't being set. This causes `childprocess` to generate an exception message that reads, "Windows says that the process completed successfully, but it did not". The process that is being referred to is the attempt to obtain the file handle for the file that should be used for streaming output. It is a little confusing, because at first glance, it seems like it might be referring to the process that should be launched.

There is an open pull request (https://github.com/enkessler/childprocess/pull/175) and corresponding issue that might resolve the issue (https://github.com/enkessler/childprocess/issues/172). For now, running on 3.0 seems like a decent enough work-around.

An issue describing this problem, with detailed steps to repeat it, will be created in the `childprocess` GitHub project. It will also be brought to the attention of the Cucumber/Aruba project that is using the `childprocess` library.